Some refactoring into components
